FP-dataset-generator/pyproject.toml
[project]
name = "fp-dataset-generator"
version = "0.2.0"
description = ""
authors = [
    {name = "Gerardo Martinez",email = "gerartmartinez@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.51.0,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "openai (>=1.0.0,<2.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "rich (>=14.2.0,<15.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"